+++ protected/controllers/MasterController.php
@@ -29,7 +29,7 @@ class MasterController extends Controller
 	{
 		return array(
 			array('allow',  // allow all users to perform 'index' and 'view' actions
-				'actions'=>array('index','view','delete'),
+				'actions'=>array('index','view'),
 				'users'=>array('*'),
 			),
 			array('allow', // allow authenticated user to perform 'create' and 'update' actions
@@ -52,9 +52,13 @@ class MasterController extends Controller
 	 */
 	public function actionView($id)
 	{
-		$this->render('view',array(
-			'model'=>$this->loadModel($id),
-		));
+		if (Yii::app()->user->checkAccess('viewMaster')) {
+			$this->render('view',array(
+				'model'=>$this->loadModel($id),
+			));
+		} else {
+			 throw new CHttpException(403, 'Нет доступа');
+		}
 	}
 
 	/**
@@ -63,21 +67,22 @@ class MasterController extends Controller
 	 */
 	public function actionCreate()
 	{
-		$model=new Master;
+		if (Yii::app()->user->checkAccess('createMaster')) {
+			$model=new Master;
 
-		// Uncomment the following line if AJAX validation is needed
-		// $this->performAjaxValidation($model);
+			if(isset($_POST['Master']))
+			{
+				$model->attributes=$_POST['Master'];
+				if($model->save())
+					$this->redirect(array('view','id'=>$model->id_master));
+			}
 
-		if(isset($_POST['Master']))
-		{
-			$model->attributes=$_POST['Master'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->id_master));
+			$this->render('create',array(
+				'model'=>$model,
+			));
+		} else {
+			 throw new CHttpException(403, 'Нет доступа');
 		}
-
-		$this->render('create',array(
-			'model'=>$model,
-		));
 	}
 
 	/**
@@ -87,21 +92,22 @@ class MasterController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
-		$model=$this->loadModel($id);
+		if (Yii::app()->user->checkAccess('updateMaster')) {
+			$model=$this->loadModel($id);
 
-		// Uncomment the following line if AJAX validation is needed
-		// $this->performAjaxValidation($model);
+			if(isset($_POST['Master']))
+			{
+				$model->attributes=$_POST['Master'];
+				if($model->save())
+					$this->redirect(array('view','id'=>$model->id_master));
+			}
 
-		if(isset($_POST['Master']))
-		{
-			$model->attributes=$_POST['Master'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->id_master));
+			$this->render('update',array(
+				'model'=>$model,
+			));
+		} else {
+			 throw new CHttpException(403, 'Нет доступа');
 		}
-
-		$this->render('update',array(
-			'model'=>$model,
-		));
 	}
 
 	/**
@@ -111,50 +117,53 @@ class MasterController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$transaction = Yii::app()->db->beginTransaction();
+		if (Yii::app()->user->checkAccess('deleteMaster')) {
+			$transaction = Yii::app()->db->beginTransaction();
+			try {
+				$master = $this->loadModel($id);
+				$animals = $master->animal;
 
-		try {
-			$master = $this->loadModel($id);
-			$animals = $master->animal;
-
-			if ($animals) {
-				foreach ($animals as $animal) {
-					$privs = $animal->priv;
-					if ($privs) {
-						foreach ($privs as $priv) {
-							$priv->delete();
-						}
-					}
-					$anemneses = $animal->anemnes;
-					if ($anemneses) {
-						foreach ($anemneses as $anemnes) {
-							$cures = $anemnes->cure;
-							if ($cures) {
-								foreach ($cures as $cure) $cure->delete();
+				if ($animals) {
+					foreach ($animals as $animal) {
+						$privs = $animal->priv;
+						if ($privs) {
+							foreach ($privs as $priv) {
+								$priv->delete();
 							}
-							$recomendations = $anemnes->recomendation;
-							if ($recomendations) {
-								foreach ($recomendations as $recomendation) $recomendation->delete();
-							}
-							$anemnes->delete();
 						}
+						$anemneses = $animal->anemnes;
+						if ($anemneses) {
+							foreach ($anemneses as $anemnes) {
+								$cures = $anemnes->cure;
+								if ($cures) {
+									foreach ($cures as $cure) $cure->delete();
+								}
+								$recomendations = $anemnes->recomendation;
+								if ($recomendations) {
+									foreach ($recomendations as $recomendation) $recomendation->delete();
+								}
+								$anemnes->delete();
+							}
+						}
+						$animal->delete();
 					}
-					$animal->delete();
 				}
-			}
-			$master->delete();
-			$transaction->commit();
+				$master->delete();
+				$transaction->commit();
 
-		} catch(Exception $e) { 
-       		$transaction->rollback();
-       		$error = $e->getMessage();
-        }
+			} catch(Exception $e) { 
+	       		$transaction->rollback();
+	       		$error = $e->getMessage();
+	        }
 
-        $this->redirect(Yii::app()->createUrl('master/index'));
+	        $this->redirect(Yii::app()->createUrl('master/index'));
 
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+			if(!isset($_GET['ajax']))
+				$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		} else {
+			 throw new CHttpException(403, 'Нет доступа');
+		}
 	}
 
 	/**
@@ -162,10 +171,14 @@ class MasterController extends Controller
 	 */
 	public function actionIndex()
 	{
-		$dataProvider=new CActiveDataProvider('Master');
-		$this->render('index',array(
-			'dataProvider'=>$dataProvider,
-		));
+		if (Yii::app()->user->checkAccess('indexMaster')) {
+			$dataProvider=new CActiveDataProvider('Master');
+			$this->render('index',array(
+				'dataProvider'=>$dataProvider,
+			));
+		} else {
+			 throw new CHttpException(403, 'Нет доступа');
+		}
 	}
 
 	/**
@@ -173,14 +186,18 @@ class MasterController extends Controller
 	 */
 	public function actionAdmin()
 	{
-		$model=new Master('search');
-		$model->unsetAttributes();  // clear any default values
-		if(isset($_GET['Master']))
-			$model->attributes=$_GET['Master'];
+		if (Yii::app()->user->checkAccess('adminMaster')) {
+			$model=new Master('search');
+			$model->unsetAttributes();  // clear any default values
+			if(isset($_GET['Master']))
+				$model->attributes=$_GET['Master'];
 
-		$this->render('admin',array(
-			'model'=>$model,
-		));
+			$this->render('admin',array(
+				'model'=>$model,
+			));
+		} else {
+			 throw new CHttpException(403, 'Нет доступа');
+		}
 	}
 
 	/**
